input: atmel_s336: report key cancellation as released.
Ubuntu touch's libinput expects key press event's value to be either
released (0) or pressed (1). If it receives canceled event (255), it'll
be treated as pressed and cause erratic problem on Ubuntu touch.
(Particularly, I'll think that the key is held and never released,
even there's a released event sent later.)

This commit makes the driver send released event instead of canceled
event, which will makes libinput happy. The long-term plan for this
is making libinput understand that this can happen, and make it report
key press-release event after it has received coresponding release key.